pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ragchat"
version = "0.1.0"
description = "A RAG-based chatbot using local LLMs"
authors = [
    {name = "Bob Blainey", email = "bob@neuralground.group"}
]

dependencies = [
    # LangChain ecosystem
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-core>=0.1.0",
    "langchain-chroma>=0.0.1",
    "langchain-ollama>=0.0.1",
    "langchain-openai>=0.0.1",
    "langchain-anthropic>=0.0.1",
    "langchain-google-genai>=0.0.1",

    # Vector store
    "chromadb>=0.4.0",

    # Document processing
    "unstructured>=0.10.0",
    "pypdf>=3.17.1",
    "python-magic>=0.4.27",
    "pdf2image>=1.16.3",
    "pytesseract>=0.3.10",
    "markdown>=3.5.1",
    "docx2txt>=0.8",
    "pdfminer.six>=20221105",
    "epub2txt>=0.1.4",
    "pypandoc>=1.11",
    "nltk>=3.8.1",

    # CLI and utilities
    "tqdm>=4.65.0",
    "pydantic>=2.0",
    "typer>=0.9.0",
    "rich>=13.3.5",
    "python-dotenv>=1.0.0",
]

requires-python = ">=3.8"

[project.scripts]
ragchat = "ragchat.main:main"

[tool.hatch.metadata]
allow-direct-references = true

# No custom build hooks needed